graphics(angle/native): glDraw{Arrays,Elements}Indirect pass full Long indirect offset (ISS-547)
AngleGL31Native truncated the Long indirect byte-offset via offsetPtr(indirect.toInt)
for glDrawArraysIndirect/glDrawElementsIndirect, while the GL31 interface takes Long
and the JVM passes the full long (MemorySegment.ofAddress). Added a Long overload of
offsetPtr (NativeGlHelper, full 64-bit castLongToRawPtr, no Int round-trip) and
dropped .toInt at both call-sites — Native now matches JVM. The Int offsetPtr and its
other callers are unchanged; glDispatchComputeIndirect was already correct.

No runtime red (only manifests with a >2GB indirect-buffer offset); verified by
fidelity (Native==JVM full-long) + clean compile native/jvm/js. fix:ae68fb57
audit:PASS (independent JVM-reference comparison).

+++ sge/src/main/scalanative/sge/graphics/AngleGL31Native.scala
@@ -145,10 +145,10 @@ class AngleGL31Native extends AngleGL30Native with GL31 {
   // ─── Indirect draw ────────────────────────────────────────────────────────
 
   override def glDrawArraysIndirect(mode: PrimitiveMode, indirect: Long): Unit =
-    GL31C.glDrawArraysIndirect(mode.toInt, offsetPtr(indirect.toInt))
+    GL31C.glDrawArraysIndirect(mode.toInt, offsetPtr(indirect))
 
   override def glDrawElementsIndirect(mode: PrimitiveMode, `type`: DataType, indirect: Long): Unit =
-    GL31C.glDrawElementsIndirect(mode.toInt, `type`.toInt, offsetPtr(indirect.toInt))
+    GL31C.glDrawElementsIndirect(mode.toInt, `type`.toInt, offsetPtr(indirect))
 
   // ─── Framebuffer parameters ───────────────────────────────────────────────
